+++ src/schemathesis/models.py
@@ -13,6 +13,7 @@ from urllib.parse import urljoin
 import attr
 import requests
 import werkzeug
+from fastapi.testclient import TestClient as FastAPIClient
 from hypothesis.strategies import SearchStrategy
 
 from .checks import ALL_CHECKS
@@ -176,6 +177,25 @@ class Case:
         client = werkzeug.Client(application, WSGIResponse)
         with cookie_handler(client, self.cookies):
             return client.open(**data, **kwargs)
+
+    def call_fastapi(
+        self,
+        app: Any = None,
+        base_url: Optional[str] = "http://testserver",
+        headers: Optional[Dict[str, str]] = None,
+        **kwargs: Any,
+    ) -> requests.Response:
+
+        application = app or self.app
+        if application is None:
+            raise RuntimeError(
+                "ASGI application instance is required. "
+                "Please, set `app` argument in the schema constructor or pass it to `call_asgi`"
+            )
+
+        client = FastAPIClient(application)
+
+        return self.call(base_url=base_url, session=client, headers=headers, **kwargs)
 
     def validate_response(
         self,